refactor: lens-debate — safeHref is duplicated as the receptacle for URL-scheme volatility — sanitize imports it from render across the DOM-free boundary
Extracted safeHref + isLoadableImage into new url-policy.ts; render.ts and sanitize.ts now import from it, removing the cross-layer back-import into the DOM-free parser.

Agreed by the lowy ⇄ hickey lens debate (finding lowy-2, raised by lowy). Not pushed or merged.
